Extract fixtures in a separate class
